refactor: baseEntity 패키지이동
