+++ src/main/java/nextstep/courses/domain/Semester.java
@@ -1,6 +1,8 @@
 package nextstep.courses.domain;
 
-public class Semester extends BaseEntity{
+import nextstep.courses.utils.BaseEntity;
+
+public class Semester extends BaseEntity {
     private final Course course;
 
     public Semester(Course course) {
